refactor(surplus): extract scheduler job bodies to surplus/jobs/
Pure structural move — zero behavior change. SurplusScheduler's ~20
embedded job bodies move to a new surplus/jobs/ package, grouped as:

- jobs/gates.py    — cooldown-gated enqueue jobs (schedule_*/brainstorm
                     check), _recently_completed, and the maintenance GC
                     block extracted as _run_maintenance_gc(db)
- jobs/runners.py  — direct-run jobs (recon gather, model intelligence,
                     skill scan, github discovery, models.md synthesis,
                     db integrity check + alarm, memory extraction,
                     follow-up dispatch)
- jobs/dream.py    — the dream-cycle pair, kept together (shared
                     _heavy_workload flag with owner-checked finally)
- jobs/gitnexus.py — reindex + CLAUDE.md strip, incl. _GITNEXUS_BLOCK_RE
- jobs/context.py  — SchedulerContext Protocol (the scheduler slice job
                     bodies read; no scheduler import, no cycle)

The scheduler keeps EVERY original method name as a thin delegate, so
APScheduler job callables, runtime wiring (JobRetryRegistry bound
brainstorm_check, init/surplus.py bound run_memory_extraction), and all
7 surplus test files keep working unchanged. _strip_gitnexus_block is
re-exported from scheduler.py (test seam); _restart_safe_hourly stays
defined there (registration infra). Function-scope imports moved
verbatim — they are the tests' patch-target seam and the import-cycle
breaker. Dream/gitnexus jobs read no scheduler state (everything via
GenesisRuntime.instance()) and take no arguments.

start() untouched; trigger classes per job id byte-identical; GROUNDWORK
count unchanged (3). Verified: record_job_* calls, event emissions, and
enqueue tuples byte-identical between old bodies and new modules.
test_code_intel_index.py entrypoint-coverage path follows the gitnexus
job to its new home in the same commit; subsystem map updated.
